feat(core): add time-decay weighting for rare variants (#57)
## Motivation

Stale typos and one-off commands from months ago competed with recent
patterns
when deciding what counts as "rare". A command typed 20 times last year
but
never since could accumulate enough raw occurrences to survive
`--remove-rare`,
even though it is effectively dead. Conversely, a typo made 3 times last
week
would survive because it hit the integer threshold. Time-decay fixes
this by
down-weighting old entries so recency matters.

## Implementation information

Entries are bucketed into three age tiers before rare-variant scoring:

| Age | Weight |
|-----|--------|
| ≤ 7 days | 1.0 |
| 8 – 30 days | 0.5 |
| > 30 days | 0.1 |

`rare_variants` now sums weighted counts per normalised command and
compares
against `rare_threshold`. Because the threshold is now compared to an
`f64`
weighted sum (not an integer occurrence count), `rare_threshold` was
retyped
from `usize` to `f64` — default stays `3.0`, same boundary, now a
weighted
score. README updated to document weights and note the flag now accepts
floats.

Closes https://github.com/Automaat/zsh-clean-history/issues/39

---------

+++ src/cli_definition.rs
@@ -13,8 +13,8 @@ const DEFAULT_LOG_MAX_BYTES: u64 = 1_048_576;
 struct Cli {
     #[arg(long, default_value_t = 0.8)]
     similarity: f64,
-    #[arg(long, default_value_t = 3)]
-    rare_threshold: usize,
+    #[arg(long, default_value_t = 3.0)]
+    rare_threshold: f64,
     #[arg(long)]
     dry_run: bool,
     #[arg(long, short = 'v')]
